requirements-jira.txt
# Dependências para o CLI de integração com Jira (opcional)
requests>=2.28.0
python-dotenv>=1.0.0
